chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.302",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]
